increase alias length upto 32 chars
