+++ x/dymns/utils/dym_name.go
@@ -66,19 +66,3 @@ func IsValidSubDymName(subDymName string) bool {
 
 	return true
 }
-
-// patternValidateAlias is a regex pattern for validating Alias (partially).
-var patternValidateAlias = regexp.MustCompile(`^[a-z\d]{1,10}$`)
-
-// IsValidAlias returns true if the given string is a valid Alias.
-func IsValidAlias(alias string) bool {
-	if alias == "" {
-		return false
-	}
-
-	if len(alias) > MaxAliasLength {
-		return false
-	}
-
-	return patternValidateAlias.MatchString(alias)
-}
