+++ x/dymns/utils/dym_name_test.go
@@ -317,67 +317,3 @@ func TestIsValidSubDymName(t *testing.T) {
 		})
 	}
 }
-
-func TestIsValidAlias(t *testing.T) {
-	tests := []struct {
-		name      string
-		alias     string
-		wantValid bool
-	}{
-		{
-			name:      "pass - valid 1 char",
-			alias:     "a",
-			wantValid: true,
-		},
-		{
-			name:      "pass - valid 2 chars",
-			alias:     "aa",
-			wantValid: true,
-		},
-		{
-			name:      "pass - valid 10 chars",
-			alias:     "1234567890",
-			wantValid: true,
-		},
-		{
-			name:      "fail - not accept 11+ chars",
-			alias:     "12345678901",
-			wantValid: false,
-		},
-		{
-			name:      "fail - not accept special chars",
-			alias:     "a$a",
-			wantValid: false,
-		},
-		{
-			name:      "fail - not accept underscore",
-			alias:     "a_a",
-			wantValid: false,
-		},
-		{
-			name:      "fail - not accept dash",
-			alias:     "a-a",
-			wantValid: false,
-		},
-		{
-			name:      "fail - not accept empty",
-			alias:     "",
-			wantValid: false,
-		},
-		{
-			name:      "fail - not accept leading space",
-			alias:     " a",
-			wantValid: false,
-		},
-		{
-			name:      "fail - not accept trailing space",
-			alias:     "a ",
-			wantValid: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require.Equal(t, tt.wantValid, IsValidAlias(tt.alias))
-		})
-	}
-}
